Finally have the sticky modifier keys bug fixed
The issue was with using workspacer window manager on Windows. Using
left alt made everything messed up and alt and other modifiers would be
stuck after using them.

So the real problem was with workspacer but I did need to change to
using right alt in Emily modifiers (see the end of file)