Add Linux support for my Plover config
Luckily it works out well to easily support Linux and windows from the
same repo. This is because windows version is nested into another
directory named plover/ making plover/plover where Linux version is just
in the only plover.

And all the dictionaries are still contained in the repo and thus there
are no absolute paths which would be bad.